Move filtering test functions to FilterTestHelpers.ts
