+++ tests/Query.test.ts
@@ -6,46 +6,12 @@ import { getSettings, updateSettings } from '../src/Settings';
 import { Query } from '../src/Query';
 import { Priority, Status, Task } from '../src/Task';
 import { createTasksFromMarkdown } from './TestHelpers';
+import {
+    FilteringCase,
+    shouldSupportFiltering,
+} from './TestingTools/FilterTestHelpers';
 
 window.moment = moment;
-
-type FilteringCase = {
-    filters: Array<string>;
-    tasks: Array<string>;
-    expectedResult: Array<string>;
-};
-
-function shouldSupportFiltering(
-    filters: Array<string>,
-    allTaskLines: Array<string>,
-    expectedResult: Array<string>,
-) {
-    // Arrange
-    const query = new Query({ source: filters.join('\n') });
-
-    const tasks = allTaskLines.map(
-        (taskLine) =>
-            Task.fromLine({
-                line: taskLine,
-                sectionStart: 0,
-                sectionIndex: 0,
-                path: '',
-                precedingHeader: '',
-            }) as Task,
-    );
-
-    // Act
-    let filteredTasks = [...tasks];
-    query.filters.forEach((filter) => {
-        filteredTasks = filteredTasks.filter(filter);
-    });
-
-    // Assert
-    const filteredTaskLines = filteredTasks.map(
-        (task) => `- [ ] ${task.toString()}`,
-    );
-    expect(filteredTaskLines).toMatchObject(expectedResult);
-}
 
 describe('Query', () => {
     describe('filtering', () => {
